Move mev inspect code to mev_inspect module
