+++ testing_file.py
@@ -1,9 +1,11 @@
-from processor import Processor
-from web3.providers import base
-from inspector_uniswap import UniswapInspector
-import block
-from web3 import Web3
 import argparse
+
+from web3.providers import base
+from web3 import Web3
+
+from mev_inspect import block
+from mev_inspect.inspector_uniswap import UniswapInspector
+from mev_inspect.processor import Processor
 
 parser = argparse.ArgumentParser(description='Inspect some blocks.')
 parser.add_argument('-block_number', metavar='b', type=int, nargs='+',
